+++ src/types/authType.ts
@@ -1,0 +1,10 @@
+export interface LoginInput {
+  email: string;
+  password: string;
+}
+
+export interface SignupInput {
+  email: string;
+  password: string;
+  nickname: string;
+}
